split utils.js file
