+++ src/utils/templates.js
@@ -1,0 +1,3 @@
+export const getTemplateWithNewData = (data, template) => {
+  return data.map((item) => template(item)).join(``);
+};
